+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Return `voteId` when submitting vote instead of the transaction hash.
+
+### Changed
+
+- Removed `dvote-solidity` dependency.
+
 ## [0.0.2-alpha] - 2022-12-20
 
 ### Added